% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.r
\name{as.matrix.fmlmat}
\alias{as.matrix.fmlmat}
\title{as_fmlmat}
\usage{
\method{as.matrix}{fmlmat}(x)
}
\arguments{
\item{x}{An object of class 'fmlmat'.}
}
\description{
Convert an fmlr matrix object (i.e. one of class 'cpumat', 'gpumat', or
'mpimat') to an 'fmlmat' object.
}
\details{
The conversion does not produce a copy of the data. It is merely a shallow
S4 wrapper around the fmlr object.
}
