% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craze-package.r
\docType{package}
\name{craze-package}
\alias{craze-package}
\title{craze}
\description{
High-level, R-like interface for fmlr. This gives a more
traditional, R-like interface around fmlr functions and methods. The
package name is a play on the German word 'fimmel'.
}
\author{
Drew Schmidt
}
\keyword{package}
