% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossprod.r
\name{crossprod}
\alias{crossprod}
\alias{crossprod,fmlmat-method}
\alias{tcrossprod,fmlmat-method}
\title{crossprod}
\usage{
\S4method{crossprod}{fmlmat}(x, y = NULL)

\S4method{tcrossprod}{fmlmat}(x, y = NULL)
}
\arguments{
\item{x}{An fmlmat matrix.}

\item{y}{Either \code{NULL}, or an fmlmat matrix.}
}
\value{
A matrix of the same type as the inputs.
}
\description{
Croddproducts.
}
