% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.r
\name{diag}
\alias{diag}
\title{diag}
\usage{
\S4method{diag}{fmlmat}(x = 1, nrow, ncol, names = TRUE)
}
\arguments{
\item{x}{fmlmat vector or matrix.}

\item{nrow, ncol}{The dimension of the resulting matrix if specified.}

\item{names}{Ignored.}
}
\description{
Create a diagonal matrix from a vector, or alternatively, get the vector of
diagonal entries from a matrix.
}
