% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.r
\name{eigen}
\alias{eigen}
\alias{eigen,fmlmat-method}
\title{eigen}
\usage{
\S4method{eigen}{fmlmat}(x, symmetric, only.values = FALSE, EISPACK = FALSE)
}
\arguments{
\item{x}{An fmlmat matrix.}

\item{symmetric, EISPACK}{Ignored.}

\item{only.values}{Should only the eigenvalues be returned?}
}
\description{
Eigenvalue decomposition.
}
\details{
Symmetric matrices only.
}
