% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes.r
\docType{class}
\name{fmlmat-class}
\alias{fmlmat-class}
\title{Class fmlmat}
\description{
An S4 container for fmlr vector/matrix objects.
}
\section{Slots}{

\describe{
\item{\code{Data}}{An fmlr object.}
}}

