% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matmult.r
\name{matmult}
\alias{matmult}
\alias{\%*\%,fmlmat,fmlmat-method}
\alias{\%*\%,fmlmat,matrix-method}
\alias{\%*\%,matrix,fmlmat-method}
\title{matmult}
\usage{
\S4method{\%*\%}{fmlmat,fmlmat}(x, y)

\S4method{\%*\%}{fmlmat,matrix}(x, y)

\S4method{\%*\%}{matrix,fmlmat}(x, y)
}
\arguments{
\item{x, y}{fmlmat matrices.}
}
\value{
A matrix of the same type as the inputs.
}
\description{
Matrix multiplication.
}
