% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print-fmlmat}
\alias{print-fmlmat}
\alias{print,fmlmat-method}
\alias{show,fmlmat-method}
\title{print-fmlmat}
\usage{
\S4method{print}{fmlmat}(x, ...)

\S4method{show}{fmlmat}(object)
}
\arguments{
\item{x, object}{An fmlmat vector/matrix.}

\item{...}{Ignored.}
}
\description{
Print methods for fmlmat objects.
}
