% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd.r
\name{svd}
\alias{svd}
\alias{La.svd,fmlmat-method}
\alias{svd,fmlmat-method}
\title{SVD}
\usage{
\S4method{La.svd}{fmlmat}(x, nu = min(n, p), nv = min(n, p))

\S4method{svd}{fmlmat}(x, nu = min(n, p), nv = min(n, p), LINPACK = FALSE)
}
\arguments{
\item{x}{An fmlmat matrix.}

\item{nu, nv}{The number of left/right singular vectors to return. NOTE: currently only
supports 0 or non-zero values.}

\item{LINPACK}{Ignored.}
}
\description{
Singular value decomposition.
}
