% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.r
\name{transpose}
\alias{transpose}
\alias{t}
\alias{t,fmlmat-method}
\title{transpose}
\usage{
t(x)

\S4method{t}{fmlmat}(x)
}
\arguments{
\item{x}{An fmlmat matrix.}
}
\value{
A matrix of the same type as the inputs.
}
\description{
Transpose.
}
