% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.r
\name{as.matrix.fmlmat}
\alias{as.matrix.fmlmat}
\title{as.matrix}
\usage{
\method{as.matrix}{fmlmat}(x, ...)
}
\arguments{
\item{x}{An object of class 'fmlmat'.}

\item{...}{Ignored.}
}
\description{
Convert an 'fmlmat' object to an R object.
}
\details{
The conversion does not produce a copy of the data. It is merely a shallow
S4 wrapper around the fmlr object.
}
