% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-constructor.r
\name{as_fmlmat}
\alias{as_fmlmat}
\title{as_fmlmat}
\usage{
as_fmlmat(x)
}
\arguments{
\item{x}{An object of class 'cpumat', 'gpumat', or 'mpimat'.}
}
\description{
Convert an fmlr matrix object (i.e. one of class 'cpumat', 'gpumat', or
'mpimat') to an 'fmlmat' object.
}
\details{
The conversion does not produce a copy of the data. It is merely a shallow
S4 wrapper around the fmlr object.
}
\examples{
library(fmlr)
library(craze)

x = matrix(1:9, 3)
x_cpumat = fmlr::as_cpumat(x, copy=FALSE)

x_fml = craze::as_fmlmat(x_cpumat)
x_fml

}
