% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.r
\name{qr}
\alias{qr}
\alias{qr.Q}
\alias{qr.R}
\alias{qr.fmlmat}
\alias{qr.Q,qr_fml-method}
\alias{qr.R,qr_fml-method}
\title{QR}
\usage{
\method{qr}{fmlmat}(x, ...)

\S4method{qr.Q}{qr_fml}(qr, complete = FALSE, Dvec)

\S4method{qr.R}{qr_fml}(qr, complete = FALSE)
}
\arguments{
\item{x}{An fmlmat matrix.}

\item{qr}{A compact QR factorization, specifically of class \code{qr_fml} - the return
of the \code{qr()} function on \code{fmlmat} data.}

\item{complete, Dvec, ...}{Ignored.}
}
\description{
QR factorization.
}
