% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.r
\name{solve}
\alias{solve}
\title{solve}
\usage{
\S4method{solve}{fmlmat}(a, b, ...)
}
\arguments{
\item{a, b}{fmlmat matrices.}

\item{...}{Ignored.}
}
\description{
Solve a system of equations or invert a matrix.
}
